# Speed up wavefunction sampling